+++ infra/crediter_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/Shopify/sarama/mocks"
 	"github.com/ethereum/go-ethereum/common"
-	"gitlab.com/ConsenSys/client/fr/core-stack/core.git/services"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/core/services"
 )
 
 func TestCrediter(t *testing.T) {